requirements.txt
pydantic==1.9.0
pyfiglet== 0.8.post1
typer==0.4.1
rich==12.2.0
textual==0.1.17
textual-inputs==0.2.5
python-dateutil==2.8.2
yfinance==0.1.70
pytest==7.1.1